This is adding a way to use idiff for comparing the result, rather than relying on vmaf.
There are some image comparisons that vmaf is problematic for, in particular if we are trying to test colorspace changes, and are expecting a 1-1 result. So we want to be able to use idiff with a specified tollerance, rather than PSNR.
This also adds the option of other comparison operators too. The goal here is to be able to use the test harness to confirm that the results match what we expect.

Signed-off-by: Sam Richards <Sam.Richards@taurich.org>
Signed-off-by: Sam.Richards@taurich.org <Sam.Richards@taurich.org>